requirements.txt
###### Requirements with Version Specifiers ######
sqlalchemy
pybase64
numpy
pandas
kaggle==1.6.17
psycopg2-binary
requests
boto3
awswrangler
snowflake-connector-python
snowflake-sqlalchemy
clickhouse-connect
httplib2
oauth2client
sendgrid
tqdm
pysftp
gspread
pyarrow==18.1.0